Change API_HOSTNAME to API_BASE
This allows http connections in addition to https.

It also allows the restriction of API server paths, shorter proxied paths,
in some cases can obfuscate some mildly sensitive parts of the path. For
example:

If we only needed to proxy the following:

- https://docs.getgrist.com/api/docs/t00TDwkP4gQ3/tables/Actions/records
- https://docs.getgrist.com/api/docs/t00TDwkP4gQ3/tables/Locations/records

We could set the API_BASE to `https://docs.getgrist.com/api/docs/t00TDwkP4gQ3`.

Then we could use the following paths with the proxy:

- http://localhost:8787/tables/Actions/records
- http://localhost:8787/tables/Locations/records